Create audio format conversion
Basically it uses a library for the conversion of the audio of WAV format to any conversion using libavcodec which is one of the library of ffmpeg. FFMPEG is basically a container which is encoded in c language and supports several other libraries in it.
The following method is one of the example or so i found,there can be others. Also it is for the WAV -> MP2. Plus one point to be considered, this library support can be of 25mb. Consider this factor also